refactor(lowlevel): move signature logic to separate package & make makeDevMsgSignature public
